Extract audio player
